refactor: separate DashboardContent into its own file, so submission view states can be passed as props
